Refactor `FeedRefreshViewModel` to `FeedPresenter`, moving to a MVP architecture, where the Presenter manages models and communicates with the view by a protocol.
The views calls the Presenter methods to access domain.

There is a two way communication channel, for example the `FeedRefreshViewController` contains a strong reference to `FeedPresenter` and the Presenter has a strong reference to the view (to display UIRefreshControl view changes).

The `FeedView` abstract view needed to display images is implemented by `FeedViewAdapter`, reusing the previous adapter method converting `FeedImages` to `FeedImageCellControllers`