Decoupled `FeedImagePresenter` from domain module, removing the responsibility of managing the image loading/cancellation. Now the communication with the models is done via an adapter: `FeedImageLoadPresentationAdapter`.
It is the responsibility of the composition layer to adapt any APIs so they communicate with each other, in this case, the adapter handles the communication between `FeedImagePresenter` and `FeedImageLoader` (Presentation <-> Domain)

Also decoupled `FeedImageCellController` from the presentation layer by using a delegate. The delegate pattern is used in this case to move the responsibility of loading the image from the presentation layer to the composition layer (via an adapter). Now the view just "asks" for the image to be loaded